rustc: be extremely principled about cross builds
There were a lot of collapsed assumptions about the cross build process
in this derivation, especially with the overloaded meanings of useLLVM.
This diff splits useLLVM into a value for each of the build, host, and
target platforms. We also specify more target-specific configure flags,
including passing through the llvm-config-native binary which is built
in some cross configurations. This also allows us to clean up some hacky
NIX_LDFLAGS code.

This commit was tested with the following build configurations:

* x86_64-freebsd -> x86_64-freebsd -> x86_64-freebsd
* x86_64-freebsd -> x86_64-freebsd -> x86_64-linux
* x86_64-freebsd -> x86_64-linux -> x86_64-linux
* x86_64-freebsd -> x86_64-freebsd -> aarch64-linux
* x86_64-freebsd -> aarch64-linux -> aarch64-linux
* x86_64-linux -> x86_64-linux -> x86_64-linux
* x86_64-linux -> x86_64-linux -> aarch64-linux
* x86_64-linux -> aarch64-linux -> aarch64-linux
* x86_64-linux -> x86_64-linux -> x86_64-freebsd
* x86_64-linux -> x86_64-freebsd -> x86_64-freebsd